+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Change Log
 
 - Change the system requirement to __macOS 11 Big Sur and later__.
 - Update the window size setting to use the last window size if the width/height setting in the preferences > Window is left blank (= auto).
+- Improve VoiceOver accessibility.
 - Update the AppleScript guide in the help.
 - Change the behavior to include the last line in the calculation when specifying lines with a negative value in the Go to Line command or via AppleScript.